commands: move next code to next.rs
